Moved the old ContentView into a new file called MainPage and called it back to ContentView, as well as creating folders for repetitive files
